What's this commit about:
1. Major Update: create new branch for the PaperNest front-end structure.

2. Major Update: Rework the branch directory structure and update some
files according to it.

3. Intermediate update: Revise page.tsx that handles citations in order to
fix button routing ("Kembali" in specific citation form writing, "batal"
in choosing citation type).

4. Intermediate update: create pages to handle displaying review requests,
versions, review detail (based on review id), and view specific versions
(based on version id).

5. Adding placeholder for invitation.

6. Intermediate update: Just moved the review-requests directory inside
[workspaceid] so the page.tsx is readable.

7. Major update: Merge from v1_dev_abiyyu due to it's working tsx files and
some configs, will pull again once he's finish his progress.

8. Major update: adding versions in the code, already fix the
implementation, so the versions doesn't screw around with the rest of
the code.

9. Intermediate update: update the part of the citation, review, and
version bar in the document to make it responsive to the mockup, and
ensure the document version can never be empty due to one of the version
is linked closely to the document creation (it's a file, not a repo.
Version cannot and shouldn't be empty).

10. Minor update: finally resolve the navbar issue in the document section
of the PaperNest web app.

11. Fix previous commit issues, pull this one as the previous ones did have
some issues.

12. Forced to use mock data, since add and list is asynchronous.

13. Testing CitationsManager function and copyDocument from Page.js, in
    order to erase no. 12