What's this about:
prevent conflict work with another branch, will definitely merge this branch with it's parent branch, then delete it.

Merge with another branch because this needs to be tested.

Merge with another branch to get latest update for this web.

Add citation page, have to be fixed later.

Remove the old page.tsx in
./src/app/[workspaceid]/documents/[documentid] directory since it's
unused and instead replace with it's JavaScript counterpart.

At least the versions was working, working on progress to actually
making the page.tsx in the citations working.